Added day five and six.

+++ src/main/kotlin/years/DayI.kt
@@ -1,6 +1,6 @@
 package org.example.years
 
 interface DayI {
-    fun part1(input: String): Int
-    fun part2(input: String): Int
+    fun part1(input: String): Long
+    fun part2(input: String): Long
 }
